db/requirements.txt
psycopg2-binary>=2.9.9
redis>=5.0.1
numpy>=1.24.0
pandas>=2.1.0
faker>=20.0.0
scipy>=1.11.0
